refactor: resolve SonarCloud findings on PR #154
dup_video_sig: AVIO callbacks become Stream's static members (ChunkAvio
shape, cpp:S5008); SigCtx deletes copy/move (owns FFmpeg resources,
cpp:S3624); the packet-feeding loop extracts to feed_decoder (cpp:S3776/
S134/S924); std::array scale buffers; init-statements. dup_scan: poster
and frame stages extracted from video_perceptual_pass (cpp:S3776),
explicit lambda captures, split chained assignment, init-statements,
using enum. dup_model: VideoSig::frame_valid widens to uint32_t — the
uint8_t bitmask tripped cpp:S6022's std::byte rule.

+++ src/ui/dup_model.cpp
@@ -171,7 +171,7 @@ bool duration_close(uint64_t a_us, uint64_t b_us) noexcept
 
 bool video_sig_match(const VideoSig& a, const VideoSig& b) noexcept
 {
-    const uint8_t common = a.frame_valid & b.frame_valid;
+    const uint32_t common = a.frame_valid & b.frame_valid;
     int shared = 0;
     int matched = 0;
     for (size_t i = 0; i < DUP_VID_FRAME_POSITIONS.size(); ++i) {
